requirements.txt
requests>=2.25
jpype1>=1.3.0
jellyfish==0.6.1
click==7.0
python-dotenv>=0.17.1
